revert miner to old version (broken version)

+++ src/blockchain/smilobft/consensus/sport/smilobftcore/core_func.go
@@ -38,13 +38,13 @@ func Encode(val interface{}) ([]byte, error) {
 	return rlp.EncodeToBytes(val)
 }
 
-func toPriority(msgCode uint64, view *sport.View) int64 {
+func toPriority(msgCode uint64, view *sport.View) float32 {
 	if msgCode == msgRoundChange {
 		// For msgRoundChange, set the message priority based on its sequence
-		return -int64(view.Sequence.Uint64() * 1000)
+		return -float32(view.Sequence.Uint64() * 1000)
 	}
 	// FIXME: round will be reset as 0 while new sequence
 	// 10 * Round limits the range of message code is from 0 to 9
 	// 1000 * Sequence limits the range of round is from 0 to 99
-	return -int64(view.Sequence.Uint64()*1000 + view.Round.Uint64()*10 + uint64(msgPriority[msgCode]))
+	return -float32(view.Sequence.Uint64()*1000 + view.Round.Uint64()*10 + uint64(msgPriority[msgCode]))
 }
